zlib: Lower the default compression level
Skia and libpng are relying on the default
value "Z_DEFAULT_COMPRESSION" which is provided
by zlib.

This commit lowers that value from 6 to 3
because devices with a fast storage and
enough memory avaliable don't need such
a high compression level.

We can reach an improvement in compression
and decompression speed by doing this.

Change-Id: I5c38463d5f10cad37921638fff39acb211209214
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>